modularized eventReducers into seperate file
